+++ GhostReplay/ReplayVehicle.hpp
@@ -37,8 +37,8 @@ public:
     // Playback control
     uint64_t GetReplayProgress();
     void TogglePause(bool pause, uint64_t gameTime);
-    void ScrubBackward(uint32_t millis);
-    void ScrubForward(uint32_t millis);
+    void ScrubBackward(uint64_t millis);
+    void ScrubForward(uint64_t millis);
 
 private:
     const CScriptSettings& mSettings;
